server_udn/requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.4.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.5.3
crcmod==1.7
Cython==3.0.10
Django==5.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
google-auth==2.30.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==1.7.1
pyparsing==3.1.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
rsa==4.9
setuptools==80.0.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
